MemoryLeakReportAutomation/requirements.txt
Django==1.9 <2
numpy==1.15.1
pytz==2018.5
pandas==0.23.4
python-dateutil==2.7.3
six==1.11.0
